Add command for Composer 2
The command for Composer 2 is available now, making it possible to
use that command from your command line as well. In the future the
command `composer` (composer v1) should be replaced with the
`composer2` command, but since it's still used a lot we will leave
that as it is for now.